Fix places controller, add places serializer
