+++ app/serializers/place_serializer.rb
@@ -1,0 +1,5 @@
+class PlaceSerializer < ActiveModel::Serializer
+  attributes :id, :name, :latitude, :longitude, :address, :description, :rules
+  has_many :events
+  has_many :images, as: :imageable
+end
